pyproject.toml
[project]
name = "personal-assistant"
version = "0.1.0"
description = "Telegram bot for meeting scheduling with Google Calendar integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "telethon>=1.34.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3"
]
